issues #52 move held events to root namespace
adding router.ObserveEventsOn(TEventProcessor). This was lifted from experimental band
