+++ src/Esp.Net.Workflow/Workflow.cs
@@ -19,7 +19,6 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using Esp.Net.Disposables;
 
 // TODO threading!! it's completely unsafe atm given the introduction of IObservable
