+++ src/Esp.Net/Esp.Net.csproj
@@ -44,7 +44,7 @@
     <Compile Include="..\AssemblyInfo.Common.cs">
       <Link>Properties\AssemblyInfo.Common.cs</Link>
     </Compile>
-    <Compile Include="BaseEventProcessor.cs" />
+    <Compile Include="RouterExt.ObserveEventsOn.cs" />
     <Compile Include="CurrentThreadDispatcher.cs" />
     <Compile Include="Disposables\CollectionDisposable.cs" />
     <Compile Include="Disposables\DictionaryDisposable.cs" />
@@ -75,7 +75,7 @@
     <Compile Include="HeldEvents\IEventDescription.cs" />
     <Compile Include="HeldEvents\IEventHoldingStrategy.cs" />
     <Compile Include="HeldEvents\IHeldEventStore.cs" />
-    <Compile Include="HeldEvents\RouterExt.cs" />
+    <Compile Include="RouterExt.HeldEvents.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reactive\EventObservable.cs" />
     <Compile Include="Reactive\EventObserver.cs" />
