+++ src/Esp.Net/HeldEvents/IHeldEventStore.cs
@@ -15,8 +15,8 @@
 #endregion
 
 #if ESP_EXPERIMENTAL
-
-namespace Esp.Net.HeldEvents
+// ReSharper disable once CheckNamespace
+namespace Esp.Net
 {
     public interface IHeldEventStore
     {
