+++ src/Esp.Net/Reactive/ModelObservable.cs
@@ -16,7 +16,8 @@
 
 using System;
 
-namespace Esp.Net.Reactive
+// ReSharper disable once CheckNamespace
+namespace Esp.Net
 {
     public interface IModelObservable<out T>
     {
