+++ src/Esp.Net/Router.cs
@@ -19,7 +19,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Esp.Net.Meta;
-using Esp.Net.Reactive;
 using Esp.Net.Utils;
 
 namespace Esp.Net
